requirements.txt
Django==3.0.3
gunicorn>=20.0.0,<20.1.0
whitenoise>=5.0.0,<5.1.0
Pillow=>7.0.0,<7.1.0
django-ckeditor==5.9.0